Added functionality to support 64bit shared lib calls completely.
Also fixed a bug that had the stack pointer incorrect after a shared
lib call because an extra call (without ret) instruction was being
used.



git-svn-id: svn+ssh://bono.sdsc.edu/projects/pmac/svnrepos/pmac/trunk/X86ElfInstrumentor@2127 55925d41-9515-0410-9586-c6f35bcc715f

+++ include/RelocationTable.h
@@ -124,7 +124,7 @@ public:
     uint16_t setSymbolTable();
     uint16_t setRelocationSection();
 
-    uint32_t addRelocation(uint32_t offset, uint32_t info);
+    uint32_t addRelocation(uint64_t offset, uint64_t info);
 
     const char* briefName() { return "RelocationTable"; }
 };
